script fix

+++ scripts/delivery.py
@@ -42,58 +42,44 @@ if __name__=='__main__':
             cCode= country.alpha_3
         if cCode in doc: 
             try:
-                os.system("echo Try onboarding for " + country.alpha_3)
-                 
-                ################## Prepare the internal structure
-                os.system("rm -rf repo")
-                os.system("rm -rf temp")
-                os.system("mkdir temp")
-                os.system("echo '"+country.alpha_3 + "\n' > temp/country")
-                os.system("echo '"+doc[cCode]+ "' > temp/base64")
-                if os.system("python scripts/config.py") !=0:
-                    raise Exception("Configuration Error")
+                branches = os.popen("git ls-remote --heads").read()
+                branchName = country.alpha_3+"/onboardingRequest"
                 
-                if os.path.exists("sync"):  
-                   ###############  Transitive Trust
-                   os.system("./scripts/transitiveTrust.sh "+country.alpha_2)
-                   if os.system("python scripts/onboardingRequest.py ./transit/"+os.environ.get("ENV")+"/countries/"+country.alpha_2) != 0:
-                          raise Exception("Onboarding Request failed.")
+                if not branchName in branches:
+                    os.system("echo Try onboarding for " + country.alpha_3)
+                    
+                    ################## Prepare the internal structure
+                    os.system("rm -rf repo")
+                    os.system("rm -rf temp")
+                    os.system("mkdir temp")
+                    os.system("echo '"+country.alpha_3 + "\n' > temp/country")
+                    os.system("echo '"+doc[cCode]+ "' > temp/base64")
+                    if os.system("python scripts/config.py") !=0:
+                        raise Exception("Configuration Error")
+                    
+                    if os.path.exists("sync"):  
+                    ###############  Transitive Trust
+                    os.system("./scripts/transitiveTrust.sh "+country.alpha_2)
+                    if os.system("python scripts/onboardingRequest.py ./transit/"+os.environ.get("ENV")+"/countries/"+country.alpha_2) != 0:
+                            raise Exception("Onboarding Request failed.")
+                    else:
+                        try:       
+                            if os.system("python scripts/repo.py") != 0:
+                                raise Exception("Repository Cloning failed.")
+                            
+                            os.system("./scripts/verify.sh 1> /dev/null")
+                            
+                            if os.system("python scripts/onboardingRequest.py repo") != 0:
+                            raise Exception("Onboarding Request failed.")
+                        except Exception as Error:
+                            os.system("echo 'Error occoured for onboarding request " + country.alpha_3 +": "+str(Error)+"'") 
+                    
+                    ######### Create PR 
+                    os.system("./scripts/createPR.sh "+country.alpha_3)
+                
+                    os.system("git checkout main > /dev/null 2>&1")
+                    os.system("git reset --hard && git clean -f -d > /dev/null 2>&1")
                 else:
-                    try:       
-                        if os.system("python scripts/repo.py") != 0:
-                              raise Exception("Repository Cloning failed.")
-                          
-                        os.system("./scripts/verify.sh 1> /dev/null")
-                        
-                        if os.system("python scripts/onboardingRequest.py repo") != 0:
-                          raise Exception("Onboarding Request failed.")
-                    except Exception as Error:
-                        os.system("echo 'Error occoured for onboarding request " + country.alpha_3 +": "+str(Error)+"'") 
-                
-            
-                
-                
-                ######### Create PR
-                
-                os.system("./scripts/createPR.sh "+country.alpha_3)
-                
-                #####################
-                if os.path.exists(country.alpha_3+"/onboarding/UP"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/UP")
-                if os.path.exists(country.alpha_3+"/onboarding/TLS"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/TLS")
-                if os.path.exists(country.alpha_3+"/onboarding/SCA"):
-                  os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/SCA")
-                if os.path.exists(country.alpha_3+"/onboarding/ISSUER"):
-                 os.system("./scripts/fileCheck.sh "+country.alpha_3+"/onboarding/ISSUER")
-                   
-                ####### Review section
-                
-                #os.system("./scripts/review.sh "+country.alpha_3)
-            
-                #####################
-            
-                os.system("git checkout main > /dev/null 2>&1")
-                os.system("git reset --hard && git clean -f -d > /dev/null 2>&1")
+                    os.system("Skip "+country.alpha_3 + "Branch already exist merge the branch or delete the branch.")
             except Exception as Error:
                 os.system("echo 'Error occoured for onboarding " + country.alpha_3 +": "+ str(Error)+"'")
